#18/feature: move Header component to Header folder
