+++ src/components/Header/Header.js
@@ -7,7 +7,7 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import { Link } from 'gatsby';
-import { blackColor, greyColor, whiteColor } from '../scss/colors.scss';
+import { blackColor, greyColor, whiteColor } from '../../scss/colors.scss';
 
 
 const useStyles = makeStyles(theme => ({
